Move free up space options into separate page
